data-collection/pyproject.toml
[tool.poetry]
name = "data-collection"
version = "0.1.0"
description = ""
authors = ["alejandro56664-2 <jheison.morales@mercadolibre.com.co>"]
readme = "README.md"
packages = [{include = "data_collection"}]

[tool.poetry.dependencies]
python = "^3.11"
gql = {extras = ["aiohttp"], version = "^3.4.0"}
python-dotenv = "^0.21.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
